requirements.txt
# generated from manifests external_dependencies
a2wsgi>=1.10.6
apispec
apispec>=4.0.0
cerberus
contextvars
extendable-pydantic
extendable-pydantic>=1.2.0
extendable>=0.0.4
fastapi>=0.110.0
graphene
graphql_server
itsdangerous
jsondiff
marshmallow
marshmallow-objects>=2.0.0
parse-accept-language
pydantic
pydantic>=2.0.0
pyjwt
pyquerystring
python-multipart
typing-extensions
ujson
